move jvm_heap_usage_threshold setting to ADNumericSetting file
